+++ src/Translations/scope/CredentialSope.ts
@@ -1,17 +1,14 @@
 import { LanguageTag } from '@Types/AppTypes';
 
-export type TranslationDTO = Record<LanguageTag, {
-  'Language': string
+type TranslationDTO = Record<LanguageTag, {
   'Credentials': string
 }>
 
-export const R_SettingsScreen: TranslationDTO = {
+export const R_CredentialScope: TranslationDTO = {
   'en-US': {
-    'Language': 'Language',
     'Credentials': 'Credentials',
   },
   'pt-BR': {
-    'Language': 'Idioma',
     'Credentials': 'Credenciais',
   },
 };