+++ docs/ReleaseNotes.md
@@ -1,3 +1,6 @@
 // Before any commit, use the command line: tsc --noEmit
 
-# 0.1.3.0 Excalidrow design docs added
+# 0.1.3.1 Credential Implementation
+
+- Added a new Service to manipulate credential files
+- Added a new Scope and screen for credentials